+++ src/main/java/interface_adapter/execute_buy/ExecuteBuyPresenter.java
@@ -3,6 +3,7 @@ package interface_adapter.execute_buy;
 import use_case.execute_buy.ExecuteBuyOutputBoundary;
 import use_case.execute_buy.ExecuteBuyOutputData;
 import utility.ViewManager;
+import view.view_events.DialogEvent;
 import view.view_events.UpdateAssetEvent;
 
 public class ExecuteBuyPresenter implements ExecuteBuyOutputBoundary {
@@ -15,17 +16,21 @@ public class ExecuteBuyPresenter implements ExecuteBuyOutputBoundary {
                         outputData.newBalance()
                 )
         );
+        // TODO: both buy and sell should be logged in transaction history
     }
 
     @Override
     public void prepareInsufficientBalanceExceptionView() {
+        ViewManager.Instance().broadcastEvent(new DialogEvent("Failed", "You have insufficient balance to buy this stock."));
     }
 
     @Override
     public void prepareStockNotFoundExceptionView() {
+        ViewManager.Instance().broadcastEvent(new DialogEvent("Failed", "The stock you are trying to buy does not exist."));
     }
 
     @Override
     public void prepareValidationExceptionView() {
+        ViewManager.Instance().broadcastEvent(new DialogEvent("Failed", "You are not authorized to do this."));
     }
 }