+++ src/main/java/use_case/login/LoginInputData.java
@@ -1,0 +1,7 @@
+package use_case.login;
+
+public record LoginInputData(
+        String username,
+        String password
+) {
+}